Исправил замечания: - Commit@Controller висит в воздухе. - В слое репозитория ставится другая аннотация над классом. - Добавить слой сервиса для RoleRepository, UserRepository и уже в класс контроллера инжекть через интерфейс сервиса, на прямую репозиторий в слой контроллера не инжектят. - @Transactional(readOnly = true) над классом убрать и перенести транзакции в сервис. - Удали весь закомментированный код и комментарии. - Используй или дао, или репозиторий. - private UserDbDAO userDbDAO; дао в слой контроллера инжектится через сервис - Зачем тебе 2 сущности User, сделай одну. - В сущностях добавить иквалс и хешкод. - Для UserService добавить интерфейс и его инжектить в контроллер.

+++ src/main/java/mvc/spring/security/repositories/RoleRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Role findByName(String name);
-// использую этот метод для реализации автозаполнения таблицы ролей в классе RoleInitializer
 }
